+++ platform/ui/src/contextProviders/ViewportGridProvider.tsx
@@ -409,6 +409,7 @@ export function ViewportGridProvider({ children, service }) {
     set: gridLayoutState => service.setState(gridLayoutState), // run it through the service itself since we want to publish events
     getNumViewportPanes,
     getActiveViewportOptionByKey,
+    setViewportGridSizeChanged: props => service.setViewportGridSizeChanged(props),
   };
 
   return (